tests/monorepo-03/apps/db-service/requirements.txt
flask==2.0.1
pyyaml==5.3.1
redis==3.5.3
